Add alternative register for slash commands and use in devtools
